# src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: NexusSettlement API
  version: 0.0.1
  description: High-throughput financial settlement and audit engine

paths:
  /api/v1/accounts:
    post:
      tags:
        - accounts
      operationId: createAccount
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAccountRequest"
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"

  /api/v1/accounts/{id}:
    get:
      tags:
        - accounts
      operationId: getAccount
      summary: Get account by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Account found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "404":
          description: Account not found

  /api/v1/health:
    get:
      tags:
        - health
      operationId: healthCheck
      summary: Health check endpoint
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResponse"

components:
  schemas:
    CreateAccountRequest:
      type: object
      required:
        - name
        - type
      properties:
        name:
          type: string
        type:
          type: string
          enum:
            - ASSET
            - LIABILITY
            - REVENUE

    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        type:
          type: string
        balance:
          type: number
          format: double
        createdAt:
          type: string
          format: date-time

    HealthResponse:
      type: object
      properties:
        status:
          type: string